clean actions and action tests
