+++ src/actions/mazeActions/mazeActions.ts
@@ -1,22 +1,13 @@
 import { AnyAction } from "redux";
-import { Coord, MazeInfo, Maze } from "../../models/maze";
-import { generateMaze } from "../../models/maze/initialState";
-import { ButtonProps } from "semantic-ui-react";
-import { PAUSE_VISUALIZATION } from "../menuActions/menuActions";
+import { Coord, MazeInfo } from "../../models/maze";
+
 export const CHANGE_START = "CHANGE_START";
 export const CHANGE_END = "CHANGE_END";
 export const MAKE_WALL = "MAKE_WALL";
 export const MAKE_EMPTY = "MAKE_EMPTY";
-export const MAKE_VISITED = "MAKE_VISITED";
-export const LOAD_MAZE = "LOAD_MAZE";
-export const STOP_VISUALIZATION = "STOP_VISUALIZATION";
 export const PROGRESS_BFS = "PROGRESS_BFS";
-export const UPDATE_GRID_SIZE = "UPDATE_GRID_SIZE";
 export const PROGRESS_DFS = "PROGRESS_DFS";
 export const PROGRESS_ASTAR = "PROGRESS_ASTAR";
-export const UPDATE_OPEN_SET = "UPDATE_OPEN_SET";
-export const UPDATE_CLOSED_SET = "UPDATE_CLOSED_SET";
-export const RANDOMIZE_WALLS = "RANDOMIZE_WALLS";
 
 export const handleChangeStart = (newPos: Coord): AnyAction => {
   return {
@@ -43,40 +34,6 @@ export const makeEmpty = (coord: Coord): AnyAction => {
   return {
     type: MAKE_EMPTY,
     payload: coord,
-  };
-};
-
-export const makeVisited = (coord: Coord): AnyAction => {
-  return {
-    type: MAKE_VISITED,
-    payload: coord,
-  };
-};
-
-export const loadMaze = (mazeInfo: MazeInfo): AnyAction => {
-  return {
-    type: LOAD_MAZE,
-    payload: { mazeInfo: mazeInfo },
-  };
-};
-
-export const handlePauseVisualization = (
-  _?: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-  data?: ButtonProps
-) => {
-  return {
-    type: PAUSE_VISUALIZATION,
-    payload: null,
-  };
-};
-
-export const handleStopVisualization = (
-  _?: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-  data?: ButtonProps
-) => {
-  return {
-    type: STOP_VISUALIZATION,
-    payload: null,
   };
 };
 
@@ -110,15 +67,6 @@ export const progressDFS = (
   };
 };
 
-export const updateGridSize = (cols: number, rows: number) => {
-  return {
-    type: UPDATE_GRID_SIZE,
-    payload: {
-      cols: cols,
-      rows: rows,
-    },
-  };
-};
 export const progressAstar = (
   openSet: Coord[],
   closedSet: Coord[],
@@ -133,26 +81,5 @@ export const progressAstar = (
       newMazeInfo,
       end,
     },
-  };
-};
-
-export const handleUpdateOpenSet = (openSet: Coord[]) => {
-  return {
-    type: UPDATE_OPEN_SET,
-    payload: openSet,
-  };
-};
-
-export const handleUpdateClosedSet = (closedSet: Coord[]) => {
-  return {
-    type: UPDATE_CLOSED_SET,
-    payload: closedSet,
-  };
-};
-
-export const randomizeWalls = () => {
-  return {
-    type: RANDOMIZE_WALLS,
-    payload: null,
   };
 };
